ADT-133 - Fix response finalise bug (#17)
* Remove POST /responses

* Refactor response repos
