+++ app/routes/responses.js
@@ -1,44 +1,8 @@
 const Joi = require('joi')
-const { addResponse, getResponses, saveResponseBlob, updateResponseMetadata, deleteResponses } = require('../storage/response-repo')
+const { getResponses, deleteResponses } = require('../storage/repos/responses')
+const { getFinalisedResponse, saveFinalisedResponse } = require('../storage/repos/finalised-responses')
 
 module.exports = [{
-  method: 'POST',
-  path: '/responses',
-  options: {
-    tags: ['api', 'responses'],
-    validate: {
-      payload: Joi.object({
-        document_id: Joi.string().uuid().required(),
-        llm: Joi.string().required(),
-        user_prompt: Joi.string().allow('').required(),
-        citations: Joi.array().required(),
-        response: Joi.string().required()
-      })
-    }
-  },
-  handler: async (request, h) => {
-    const document_id = request.payload.document_id
-    const llm = request.payload.llm
-    const user_prompt = request.payload.user_prompt
-    const citations = request.payload.citations
-    const response = request.payload.response
-
-    const data = {
-      document_id,
-      llm,
-      user_prompt,
-      citations,
-      response
-    }
-
-    await addResponse(
-      data
-    )
-
-    return h.response().code(201)
-  }
-},
-{
   method: 'GET',
   path: '/responses/{docId}',
   options: {
@@ -61,38 +25,14 @@ module.exports = [{
   }
 },
 {
-  method: 'POST',
-  path: '/responses/finalise',
-  handler: async (request, h) => {
-    const responses = await getResponses(request.payload.document_id)
-    const latestResponseBuffer = Buffer.from(responses[0].response)
-
-    const responseStorageOptions = {
-      latestResponseBuffer,
-      projectName: request.payload.project_name,
-      documentId: request.payload.document_id
-    }
-
-    try {
-      const savedLatestResponseBuffer = await saveResponseBlob(responseStorageOptions)
-      return h.response(savedLatestResponseBuffer).code(201)
-    } catch (err) {
-      return h.response({ error: err.message }).code(500)
-    }
-  }
-},
-{
-  method: 'PUT',
+  method: 'GET',
   path: '/responses/finalise/{projectName}/{documentId}',
   handler: async (request, h) => {
-    const updateResponseMetadataOptions = {
-      projectName: request.params.projectName,
-      documentId: request.params.documentId
-    }
-
     try {
-      const status = await updateResponseMetadata(updateResponseMetadataOptions)
-      return h.response(status).code(200)
+      const { projectName, documentId } = request.params
+      const response = await getFinalisedResponse(projectName, documentId)
+
+      return h.response(response).code(200)
     } catch (err) {
       if (err.code === 'NotFound') {
         return h.response().code(404).takeover()
@@ -103,20 +43,69 @@ module.exports = [{
   }
 },
 {
-  method: 'DELETE',
-  path: '/responses/{docId}',
+  method: 'POST',
+  path: '/responses/finalise',
   options: {
-    tags: ['api', 'responses'],
     validate: {
-      params: Joi.object({
-        docId: Joi.string().uuid().required()
+      payload: Joi.object({
+        projectName: Joi.string().required(),
+        documentId: Joi.string().uuid().required()
       })
     }
   },
   handler: async (request, h) => {
-    await deleteResponses(
-      request.params.docId
-    )
+    const { projectName, documentId } = request.payload
+
+    let responses
+
+    try {
+      responses = await getResponses(documentId)
+    } catch (err) {
+      console.error('Error getting responses:', err)
+      throw err
+    }
+
+    const { response } = responses[0]
+
+    await saveFinalisedResponse(projectName, documentId, response)
+
+    return h.response().code(201)
+  }
+},
+{
+  method: 'PUT',
+  path: '/responses/finalise/{projectName}/{documentId}',
+  options: {
+    validate: {
+      payload: Joi.object({
+        response: Joi.string().required()
+      })
+    }
+  },
+  handler: async (request, h) => {
+    const { projectName, documentId } = request.params
+    const { response } = request.payload
+
+    await saveFinalisedResponse(projectName, documentId, response)
+
+    return h.response().code(200)
+  }
+},
+{
+  method: 'DELETE',
+  path: '/responses/{documentId}',
+  options: {
+    tags: ['api', 'responses'],
+    validate: {
+      params: Joi.object({
+        documentId: Joi.string().uuid().required()
+      })
+    }
+  },
+  handler: async (request, h) => {
+    const { documentId } = request.params
+
+    await deleteResponses(documentId)
 
     return h.response().code(200)
   }
