+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "coreai-mcu-responses",
-  "version": "0.1.15",
+  "version": "0.1.16",
   "description": "",
   "homepage": "github.com?owner=defra&repo=coreai-mcu-responses&organization=defra",
   "main": "app/index.js",
